Replace dependency-versions-check with requireUpperBoundDeps
This replaces maven-dependency-versions-check-plugin with the
requireUpperBoundDeps Enforcer rule. In my testing, the rule caught
everything the plugin caught, plus some situations it did not.

The plugin is very CPU intensive and is often the slowest of the
checkers, which increases the (non-test) build time noticeably. We
attempted to improve that by making the plugin multithreaded, but it
ends up crashing, likely due to non-threadsafe code in Maven itself.
The rule is substantially faster.