Fix the way locale information is passed to Surefire
This moves primarily the `user.language` and `user.region` from
`systemPropertyVariables` to explicit property setting on the JVM. These
two methods to set them are supposed to be the same, but passing via
system properties does not actually work. Which is annoying, because
when the are set in both ways, Surefire complains about it:

```
[WARN] The system property user.language is configured twice! The property appears in <argLine/> and any of <systemPropertyVariables/>, <systemProperties/> or user property.
[WARN] The system property user.region is configured twice! The property appears in <argLine/> and any of <systemPropertyVariables/>, <systemProperties/> or user property.
```

It seems that the `user.timezone` actually works as expected, but is
moved along with the other two for consistency.