Always declare empty argLine
The `argLine` property is used by Jacoco to store the JVM commands which
enable its agent. It is initialized by the Jacoco plugin, then used in
the Surefire configuration. But when Jacoco is disabled (e.g. using the
`air.check.skip-jacoco` property), the `argLine` property is not
initialized, and if we don't take additional measures, it will appear as
literal `${argLine}` in Surefire's JVM arguments line, which may cause
trouble with some tools. The way it was handled previously was that
there was an additional profile, which was supposed to be activated when
Jacoco was disabled, and which initialized the `argLine` property to
empty, allowing Maven to properly interpolate it in the Surefire
configuration.

There was one problem with this solution, though - these are system
properties, not Maven properties, and profile activation is done before
property values are resolved. So if we have a project which disables
Jacoco explicitly in its POM (by overriding the property like this:
`<air.check.skip-jacoco>true</air.check.skip-jacoco>`), this did not
change the set of active profiles. This means that the Jacoco plugin was
still configured - though disabled - and the prifile to initialize the
`argLine` property was still not active. Thus, when the Surefire
configuration was evaluated, the `argLine` property was uninitialized.

The fix in this commit unconditionally initializes `argLine` to empty.
This way if Jacoco is enabled, the Jacoco plugin will overwrite this
property; otherwise it will remain empty. But for this to work we have
to change the Surefire configuration to use late evaluation for
`argLine` to prevent its evaluation before it's reinitialized by the
Jacoco plugin.