Suppress Checkstyle within text-blocks
Some checks like RegexpMultiline are not AST aware and hence cannot
differentiate between content inside a text-block or outside of it. This
is a short term workaround which uses the """ text as a marker to let
Checkstyle suppress any violations between them.

The longer term fix is to migrate away from non-AST aware checks like
RegexpMultiline and RegexpSingleline to something else.