Fix running timezone-sensitive tests in IntelliJ IDEA
Some time ago, some properties were moved to `argLine` because Surefire
incorrectly passes the `user.language` and `user.region` system
properties to the tests. At that time, the `user.timezone` also got
along for the ride, but it turns out, that that latter property *is*
passed correctly. So while it works for Maven and Surefire, moving it
into `argLine` causes IntelliJ IDEA to not pick it up, which in turn
means that some tests are not working properly when executed inside the
IDE. We can move it safely back, as it improves quality of life.

This reverts parts of commit d4630aa6547a64c7112f2bf133138d7d83b6d1a8.